Move ScanThrownTypes out of InterruptedExceptionSwallowed, so it can be more easily reused.
Add some separate tests for it. Add support for the rethrowing-caught-variable behaviour.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=307610905
