+++ core/src/main/java/com/google/errorprone/bugpatterns/InterruptedExceptionSwallowed.java
@@ -16,7 +16,6 @@
 
 package com.google.errorprone.bugpatterns;
 
-import static com.google.common.base.Verify.verify;
 import static com.google.common.collect.Iterables.getLast;
 import static com.google.errorprone.BugPattern.SeverityLevel.WARNING;
 import static com.google.errorprone.matchers.Description.NO_MATCH;
@@ -28,7 +27,6 @@ import static com.google.errorprone.matchers.Matchers.methodHasVisibility;
 import static com.google.errorprone.matchers.Matchers.methodIsNamed;
 import static com.google.errorprone.matchers.Matchers.methodReturns;
 import static com.google.errorprone.matchers.MethodVisibility.Visibility.PUBLIC;
-import static com.google.errorprone.util.ASTHelpers.getSymbol;
 import static com.google.errorprone.util.ASTHelpers.getType;
 import static com.google.errorprone.util.ASTHelpers.isSameType;
 import static com.google.errorprone.util.ASTHelpers.isSubtype;
@@ -47,34 +45,21 @@ import com.google.errorprone.fixes.SuggestedFix;
 import com.google.errorprone.fixes.SuggestedFixes;
 import com.google.errorprone.matchers.Description;
 import com.google.errorprone.matchers.Matcher;
-import com.google.errorprone.suppliers.Supplier;
 import com.google.errorprone.suppliers.Suppliers;
 import com.google.errorprone.util.ASTHelpers;
+import com.google.errorprone.util.ASTHelpers.ScanThrownTypes;
 import com.sun.source.tree.BlockTree;
 import com.sun.source.tree.CatchTree;
-import com.sun.source.tree.ClassTree;
 import com.sun.source.tree.InstanceOfTree;
-import com.sun.source.tree.LambdaExpressionTree;
-import com.sun.source.tree.MethodInvocationTree;
 import com.sun.source.tree.MethodTree;
-import com.sun.source.tree.NewClassTree;
 import com.sun.source.tree.StatementTree;
-import com.sun.source.tree.ThrowTree;
 import com.sun.source.tree.TryTree;
-import com.sun.source.tree.VariableTree;
 import com.sun.source.util.TreeScanner;
-import com.sun.tools.javac.code.Symbol;
-import com.sun.tools.javac.code.Symbol.ClassSymbol;
-import com.sun.tools.javac.code.Symbol.MethodSymbol;
 import com.sun.tools.javac.code.Type;
 import com.sun.tools.javac.code.Type.UnionClassType;
-import com.sun.tools.javac.code.Types;
 import com.sun.tools.javac.tree.JCTree;
-import com.sun.tools.javac.util.Name;
-import java.util.ArrayDeque;
 import java.util.HashSet;
 import java.util.List;
-import java.util.Optional;
 import java.util.Set;
 import java.util.stream.Stream;
 import javax.annotation.Nullable;
@@ -117,7 +102,7 @@ public final class InterruptedExceptionSwallowed extends BugChecker
     if (tree.getThrows().stream().anyMatch(t -> isSubtype(getType(t), interrupted, state))) {
       return NO_MATCH;
     }
-    ImmutableSet<Type> thrownExceptions = getThrownExceptions(tree.getBody(), state);
+    ImmutableSet<Type> thrownExceptions = ASTHelpers.getThrownExceptions(tree.getBody(), state);
     // Bail out if none of the exceptions thrown are subtypes of InterruptedException.
     if (thrownExceptions.stream().noneMatch(t -> isSubtype(t, interrupted, state))) {
       return NO_MATCH;
@@ -213,13 +198,6 @@ public final class InterruptedExceptionSwallowed extends BugChecker
         }.scan(block, null));
   }
 
-  /** Returns the exceptions thrown by {@code blockTree}. */
-  private static ImmutableSet<Type> getThrownExceptions(BlockTree blockTree, VisitorState state) {
-    ScanThrownTypes scanner = new ScanThrownTypes(state);
-    scanner.scan(blockTree, null);
-    return ImmutableSet.copyOf(scanner.getThrownTypes());
-  }
-
   /**
    * Returns the exceptions that need to be handled by {@code tryTree}'s catch blocks, or be
    * propagated out.
@@ -229,133 +207,6 @@ public final class InterruptedExceptionSwallowed extends BugChecker
     scanner.scanResources(tryTree);
     scanner.scan(tryTree.getBlock(), null);
     return ImmutableSet.copyOf(scanner.getThrownTypes());
-  }
-
-  static final class ScanThrownTypes extends TreeScanner<Void, Void> {
-    boolean inResources = false;
-    ArrayDeque<Set<Type>> thrownTypes = new ArrayDeque<>();
-
-    private final VisitorState state;
-    private final Types types;
-
-    ScanThrownTypes(VisitorState state) {
-      this.state = state;
-      this.types = state.getTypes();
-      thrownTypes.push(new HashSet<>());
-    }
-
-    private Set<Type> getThrownTypes() {
-      return thrownTypes.peek();
-    }
-
-    @Override
-    public Void visitMethodInvocation(MethodInvocationTree invocation, Void unused) {
-      MethodSymbol symbol = getSymbol(invocation);
-      if (symbol != null) {
-        getThrownTypes().addAll(symbol.getThrownTypes());
-      }
-      return super.visitMethodInvocation(invocation, null);
-    }
-
-    @Override
-    public Void visitTry(TryTree tree, Void unused) {
-      thrownTypes.push(new HashSet<>());
-      scanResources(tree);
-      scan(tree.getBlock(), null);
-      // Make two passes over the `catch` blocks: once to remove caught exceptions, and once to
-      // add thrown ones. We can't do this in one step as an exception could be caught but later
-      // thrown.
-      for (CatchTree catchTree : tree.getCatches()) {
-        Type type = getType(catchTree.getParameter());
-        for (Type unionMember : extractTypes(type)) {
-          getThrownTypes().removeIf(thrownType -> types.isAssignable(thrownType, unionMember));
-        }
-      }
-      for (CatchTree catchTree : tree.getCatches()) {
-        scan(catchTree.getBlock(), null);
-      }
-      scan(tree.getFinallyBlock(), null);
-      Set<Type> fromBlock = thrownTypes.pop();
-      getThrownTypes().addAll(fromBlock);
-      return null;
-    }
-
-    public void scanResources(TryTree tree) {
-      inResources = true;
-      scan(tree.getResources(), null);
-      inResources = false;
-    }
-
-    @Override
-    public Void visitThrow(ThrowTree tree, Void unused) {
-      getThrownTypes().addAll(extractTypes(getType(tree.getExpression())));
-      return super.visitThrow(tree, null);
-    }
-
-    @Override
-    public Void visitNewClass(NewClassTree tree, Void unused) {
-      MethodSymbol symbol = getSymbol(tree);
-      if (symbol != null) {
-        getThrownTypes().addAll(symbol.getThrownTypes());
-      }
-      return super.visitNewClass(tree, null);
-    }
-
-    @Override
-    public Void visitVariable(VariableTree tree, Void unused) {
-      if (inResources) {
-        Symbol symbol = getSymbol(tree.getType());
-        if (symbol instanceof ClassSymbol) {
-          getCloseMethod((ClassSymbol) symbol, state)
-              .ifPresent(methodSymbol -> getThrownTypes().addAll(methodSymbol.getThrownTypes()));
-        }
-      }
-      return super.visitVariable(tree, null);
-    }
-
-    // We don't need to account for anything thrown by declarations.
-    @Override
-    public Void visitLambdaExpression(LambdaExpressionTree tree, Void unused) {
-      return null;
-    }
-
-    @Override
-    public Void visitClass(ClassTree tree, Void unused) {
-      return null;
-    }
-
-    @Override
-    public Void visitMethod(MethodTree tree, Void unused) {
-      return null;
-    }
-  }
-
-  private static final Supplier<Type> AUTOCLOSEABLE =
-      Suppliers.typeFromString("java.lang.AutoCloseable");
-  private static final Supplier<Name> CLOSE = VisitorState.memoize(state -> state.getName("close"));
-
-  private static Optional<MethodSymbol> getCloseMethod(ClassSymbol symbol, VisitorState state) {
-    Types types = state.getTypes();
-    if (!types.isAssignable(symbol.type, AUTOCLOSEABLE.get(state))) {
-      return Optional.empty();
-    }
-    Type voidType = state.getSymtab().voidType;
-    Optional<MethodSymbol> declaredCloseMethod =
-        ASTHelpers.matchingMethods(
-                CLOSE.get(state),
-                s ->
-                    !s.isConstructor()
-                        && s.params.isEmpty()
-                        && types.isSameType(s.getReturnType(), voidType),
-                symbol.type,
-                types)
-            .findFirst();
-    verify(
-        declaredCloseMethod.isPresent(),
-        "%s implements AutoCloseable but no method named close() exists, even inherited",
-        symbol);
-
-    return declaredCloseMethod;
   }
 
   private static ImmutableList<Type> extractTypes(@Nullable Type type) {
